+++ app/src/main/java/com/example/compress/util/Embed.java
@@ -12,7 +12,7 @@ public class Embed {
      * @param code
      * @return
      */
-    public static int[] embed(int a, int b, int[] bitmap, int code) {
+    public static double[] embed(double a, double b, double[] bitmap, double code) {
         if (a == b) {
             if (code == 1) {
                 for (int i = 0; i < bitmap.length; i++) {
@@ -26,18 +26,18 @@ public class Embed {
         }
         if (code == 1) {
             if (a < b) {
-                int temp = a;
+                double temp = a;
                 a = b;
                 b = temp;
             }
         } else if (code == 0) {
             if (a > b) {
-                int temp = a;
+                double temp = a;
                 a = b;
                 b = temp;
             }
         }
-        int[] group4 = {a, b, bitmap[0], bitmap[1]};
+        double[] group4 = {a, b, bitmap[0], bitmap[1]};
         return group4;
     }
 }
