+++ app/src/main/java/com/example/compress/util/My_rand.java
@@ -21,12 +21,12 @@ public class My_rand {
      * @param key2
      * @return
      */
-    public static int[] my_rand(int[] bitmap, double key1, double key2) {
+    public static double[] my_rand(double[] bitmap, double key1, double key2) {
         int len = bitmap.length;
-        int[] bitmap2 = new int[len];
-        int[] rand_order = Chaotic.chaotic_maping_order(key1, key2, 16);
+        double[] bitmap2 = new double[len];
+        double[] rand_order = Chaotic.chaotic_maping_order(key1, key2, 16);
         for (int i = 0; i < len; i++) {
-            bitmap2[i] = bitmap[rand_order[i]];
+            bitmap2[i] = bitmap[(int)rand_order[i]];
         }
         return bitmap2;
     }
